Use HTTPS for d3
In order for scripts to be loaded when the notebook server is served over HTTPS, scripts need to fetched from HTTPS as well. Chrome for instance will actively block HTTP scripts when a domain is served over HTTPS.

Note that the cdnjs URL is the "canonical" URL that http://d3js.org/ lists.